requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
psycopg2-binary==2.9.7
python-dotenv==1.0.0
streamlit==1.28.0
pandas==2.0.3
plotly==5.15.0
gunicorn==21.2.0